% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_load_atlas_data.R
\name{read_dataverse_atlas}
\alias{read_dataverse_atlas}
\title{Read Dataset Files from Directory}
\usage{
read_dataverse_atlas(digits, years, dir, workers = 10)
}
\arguments{
\item{digits}{An integer indicating the level of digit classification of the dataset (2, 4, or 6).}

\item{years}{A vector of integers indicating the years of interest.}

\item{dir}{A character string specifying the directory path where datasets are located.}

\item{workers}{An integer specifying the number of workers to use for parallel processing (default is 10).}
}
\value{
A data frame containing combined data from all matched .dta files.
}
\description{
Reads .dta files matching specified digit level and years from a directory into a single data frame.
It performs an error check to ensure there are files matching the query before attempting to read them.
}
\examples{
read_dataverse_atlas(digits = 4, years = 2010:2021, dir = "01_raw_data/aec")
}
